requirements.txt
moviepy>=1.0.3
Flask>=3.0.0
gTTS>=2.5.0
Pillow==10.4.0
numpy>=1.24.3
pydub>=0.25.1
